Fix the darwin build, install fuse3 in CI, document the mount
The darwin cross-compile failed, and both errors were mine:

- syscall.Mknod on darwin takes (string, uint32, int) like Linux; the
  uint16/int32 conversions I guessed at were wrong.
- Go's darwin syscall package does not export MNT_FORCE, so the detach
  flag now comes from x/sys/unix instead of being hardcoded.

Verified by cross-compiling every release target locally rather than
letting CI find it: linux, darwin and windows on amd64 and arm64.

CI installs fuse3 so the mount tests actually run there. A mount that
only works on a developer's machine is not a tested mount, and a silently
skipped test is exactly the deceptive green this repo's conventions warn
about.

README and CLAUDE.md now describe the mount: why replacing file tools
cannot be enough, why the mount point defaults to the remote's own path,
the coherence window on caching, and the failure modes worth knowing
(contagious dead mounts, O_APPEND, cross-platform open flags, and the
go-fuse checksum quirk in the org proxy).

+++ agent/fsstat_darwin.go
@@ -52,7 +52,7 @@ func statfs(path string) (*fswire.Statfs, error) {
 
 // mknod creates a device or FIFO node.
 func mknod(path string, mode, dev uint32) error {
-	return syscall.Mknod(path, uint16(mode), int32(dev))
+	return syscall.Mknod(path, mode, int(dev))
 }
 
 // utimes sets access and modification times, leaving unset ones untouched
